- Ported ZDoom revisions 2480, 2482, 2483 and 2485 in a new branch. - Solved conflicts and repaired compilation errors as well as I could. - Marked most troublesome spots with POLYBSP_FIXME.
This compiles but does NOT work! Some gl_bsp.cpp code was entirely commented out, and the game crashes during level setup. The main reason for making this a branch is to hopefully make it easier to fold back into the main trunk once the GL code has been properly adapted to ZDoom's new system.

git-svn-id: http://mancubus.net/svn/hosted/gzdoom/branches/polybsp@868 b0f79afe-0144-0410-b225-9a4edf0717df